+++ src/functions.php
@@ -84,8 +84,6 @@ namespace Rubix\ML
         foreach ($iterator as $element) {
             return $element;
         }
-
-        return null;
     }
 
     /**